requirements.txt
# Core LangChain ecosystem
langchain==0.3.27
langchain-core==0.3.76
langchain-community==0.3.29
langchain-openai==0.3.33
langsmith==0.4.30
langchain_experimental==0.3.4
# OpenAI SDK
openai==1.108.2

# PDF handling
pypdf==6.1.0
# (PyPDF2 is legacy; you can remove it unless you really want it)

# Utilities
python-dotenv==1.1.1   # for .env file
tiktoken==0.11.0       # tokenization support
numpy==2.3.3           # required by LangChain/ML
pandas==2.3.2          # structured data handling

# NLP & parsing (optional, but useful for resumes / text cleaning)
nltk==3.9.1
regex==2025.9.18

# Web & requests
requests==2.32.5
urllib3==2.5.0

# JSON / serialization
orjson==3.11.3

# Table formatting (optional but nice for debugging / printing)
tabulate==0.9.0
scikit-learn==1.7.2
python-docx==1.2.0
newspaper3k==0.2.8
PyPDF2==3.0.1
lxml
lxml_html_clean
faiss-cpu
streamlit==1.49.1